renamed utils to update per suggestion
