+++ src/zenml/models/v2/core/component.py
@@ -37,7 +37,7 @@ from zenml.models.v2.base.scoped import (
     ShareableResponseBody,
     ShareableResponseMetadata,
 )
-from zenml.models.v2.base.utils import update_model
+from zenml.models.v2.base.update import update_model
 from zenml.utils import secret_utils
 
 if TYPE_CHECKING:
