+++ src/zenml/models/v2/core/flavor.py
@@ -27,7 +27,7 @@ from zenml.models.v2.base.base import (
     BaseResponseMetadata,
 )
 from zenml.models.v2.base.scoped import WorkspaceScopedFilter
-from zenml.models.v2.base.utils import update_model
+from zenml.models.v2.base.update import update_model
 
 if TYPE_CHECKING:
     from zenml.models import (
