+++ src/zenml/models/v2/core/service_connector.py
@@ -29,7 +29,7 @@ from zenml.models.v2.base.scoped import (
     ShareableResponseBody,
     ShareableResponseMetadata,
 )
-from zenml.models.v2.base.utils import update_model
+from zenml.models.v2.base.update import update_model
 from zenml.models.v2.misc.service_connector_type import (
     ServiceConnectorTypeModel,
 )
