+++ src/zenml/models/v2/core/team.py
@@ -26,7 +26,7 @@ from zenml.models.v2.base.base import (
     BaseResponseMetadata,
 )
 from zenml.models.v2.base.filter import BaseFilter
-from zenml.models.v2.base.utils import update_model
+from zenml.models.v2.base.update import update_model
 from zenml.models.v2.core.user import UserResponse
 
 # ------------------ Request Model ------------------
