+++ src/zenml/models/v2/core/workspace.py
@@ -25,7 +25,7 @@ from zenml.models.v2.base.base import (
     BaseResponseMetadata,
 )
 from zenml.models.v2.base.filter import BaseFilter
-from zenml.models.v2.base.utils import update_model
+from zenml.models.v2.base.update import update_model
 
 # ------------------ Request Model ------------------
